+++ app/Console/Kernel.php
@@ -27,8 +27,7 @@ class Kernel extends ConsoleKernel
      */
     protected function schedule(Schedule $schedule)
     {
-        // $schedule->command('inspire')
-        //          ->hourly();
+        $schedule->command('horizon:snapshot')->everyFiveMinutes();
     }
 
     /**